Drop sense pin usage.
Some mainboards don't implement sense pin properly and have other
signals connected to that pin. This can lead to invalid output voltage
reading. In the worst case sense could be bound to ground, which would
report 0V to the controller all the time, which would try to ramp up
the voltage to the maximum of about 4.4V. This could potentially kill
the CPU, especially low voltage K6-II or similar.

Due to this issue, the feedback voltage is now read directly at the
vcore output. This could potentially lead to slightly lower output
voltage at the CPU, since we would not read the sense from there, but
it is much safer and would work on more mainboards.